+++ src/api/missionApi.ts
@@ -1,5 +1,5 @@
 import apiClient from './httpClient';
-import { MissionHistoryResponse } from '../types/achievemnet';
+import { MissionHistoryResponse } from '../types/achievement';
 
 export const getMissionHistory = async (): Promise<MissionHistoryResponse> => {
   const { data } = await apiClient.get<MissionHistoryResponse>(
