+++ src/pages/achievement/mission/MockData.tsx
@@ -1,0 +1,53 @@
+// 목업데이터
+export const mockCompletedMissions = {
+  missions: [
+    {
+      mission_check_id: 301,
+      mission_id: 1,
+      title: '오늘 5,000보 걷기',
+      category: 'STEP',
+      period_type: 'DAILY',
+      period_start: '2025-12-18',
+      period_end: '2025-12-18',
+      goal_value: 5000,
+      progress_value: 5000,
+      completed_at: '2025-12-18T18:42:10',
+    },
+    {
+      mission_check_id: 298,
+      mission_id: 11,
+      title: '하루 세 끼 식단 기록',
+      category: 'MEAL',
+      period_type: 'DAILY',
+      period_start: '2025-12-17',
+      period_end: '2025-12-17',
+      goal_value: 3,
+      progress_value: 3,
+      completed_at: '2025-12-18T21:10:33',
+    },
+    {
+      mission_check_id: 287,
+      mission_id: 5,
+      title: '주간 35,000보 걷기',
+      category: 'STEP',
+      period_type: 'WEEKLY',
+      period_start: '2025-12-09',
+      period_end: '2025-12-15',
+      goal_value: 35000,
+      progress_value: 32540,
+      completed_at: '2025-12-15T23:55:01',
+    },
+    {
+      mission_check_id: 241,
+      mission_id: 100,
+      title: '월간 150,000보 걷기',
+      category: 'STEP',
+      period_type: 'MONTHLY',
+      period_start: '2025-11-01',
+      period_end: '2025-11-30',
+      goal_value: 150000,
+      progress_value: 108230,
+      completed_at: '2025-11-30T22:10:00',
+    },
+  ],
+};
